requirements.txt
Django==1.7.1
django-angular==0.7.9
django-filter==0.8
djangorestframework==2.4.4
psycopg2==2.5.4
six==1.8.0
wsgiref==0.1.2
